final version

+++ libft/ft_strlen.c
@@ -3,19 +3,18 @@
 /*                                                        ::::::::            */
 /*   ft_strlen.c                                        :+:    :+:            */
 /*                                                     +:+                    */
-/*   By: pderksen <pderksen@student.codam.nl>         +#+                     */
+/*   By: sde-quai <sde-quai@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
-/*   Created: 2021/10/28 12:24:13 by pderksen      #+#    #+#                 */
-/*   Updated: 2021/10/28 12:24:14 by pderksen      ########   odam.nl         */
+/*   Created: 2021/12/02 12:56:58 by sde-quai      #+#    #+#                 */
+/*   Updated: 2021/12/02 12:56:58 by sde-quai      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <unistd.h>
-#include <stdio.h>
+#include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
